backend/pyproject.toml
[tool.poetry]
name = "cotai-backend"
version = "0.1.0"
description = "Backend para o sistema CotAi de gestão de cotações para licitações"
authors = ["CotAi Team <contato@cotai.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
pydantic = {extras = ["email"], version = "^2.0.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.11.0"
psycopg = {extras = ["binary"], version = "^3.1.10"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "^4.6.0"
pymongo = "^4.4.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
tenacity = "^8.2.2"
pydantic-settings = "^2.0.0"
pyotp = "^2.9.0"
qrcode = "^7.4.2"
Pillow = "^10.0.0"
fastapi-limiter = "^0.1.5"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.4.0"
pre-commit = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
